chore: move some stuff to utils
