+++ evm-adapters/src/sputnik/cheatcodes/cheatcode_handler.rs
@@ -41,7 +41,7 @@ use crate::sputnik::cheatcodes::{
 };
 use once_cell::sync::Lazy;
 
-use crate::sputnik::utils::convert_log;
+use crate::sputnik::utils::{convert_log, evm_error, revert_return_evm, ExpectRevertReturn};
 use ethers::abi::Tokenize;
 
 // This is now getting us the right hash? Also tried [..20]
@@ -56,37 +56,6 @@ pub static CHEATCODE_ADDRESS: Lazy<Address> = Lazy::new(|| {
 pub static CONSOLE_ADDRESS: Lazy<Address> = Lazy::new(|| {
     Address::from_slice(&hex::decode("000000000000000000636F6e736F6c652e6c6f67").unwrap())
 });
-
-/// Wrapper around both return types for expectRevert in call or create
-enum ExpectRevertReturn {
-    Call(Capture<(ExitReason, Vec<u8>), Infallible>),
-    Create(Capture<(ExitReason, Option<H160>, Vec<u8>), Infallible>),
-}
-
-impl ExpectRevertReturn {
-    pub fn into_call_inner(self) -> Capture<(ExitReason, Vec<u8>), Infallible> {
-        match self {
-            ExpectRevertReturn::Call(inner) => inner,
-            _ => panic!("tried to get call response inner from a create"),
-        }
-    }
-    pub fn into_create_inner(self) -> Capture<(ExitReason, Option<H160>, Vec<u8>), Infallible> {
-        match self {
-            ExpectRevertReturn::Create(inner) => inner,
-            _ => panic!("tried to get create response inner from a call"),
-        }
-    }
-
-    pub fn is_call(&self) -> bool {
-        matches!(self, ExpectRevertReturn::Call(..))
-    }
-}
-
-/// For certain cheatcodes, we may internally change the status of the call, i.e. in
-/// `expectRevert`. Solidity will see a successful call and attempt to abi.decode for the called
-/// function. Therefore, we need to populate the return with dummy bytes such that the decode
-/// doesn't fail
-pub static DUMMY_OUTPUT: [u8; 320] = [0u8; 320];
 
 /// Hooks on live EVM execution and forwards everything else to a Sputnik [`Handler`].
 ///
@@ -348,51 +317,6 @@ impl<'a, 'b, B: Backend, P: PrecompileSet>
         ]);
 
         evm
-    }
-}
-
-// helper for creating an exit type
-fn evm_error(retdata: &str) -> Capture<(ExitReason, Vec<u8>), Infallible> {
-    Capture::Exit((
-        ExitReason::Revert(ExitRevert::Reverted),
-        ethers::abi::encode(&[Token::String(retdata.to_owned())]),
-    ))
-}
-
-// helper for creating the Expected Revert return type, based on if there was a call or a create,
-// and if there was any decoded retdata that matched the expected revert value.
-fn revert_return_evm<T: ToString>(
-    call: bool,
-    result: Option<(&[u8], &[u8])>,
-    err: impl FnOnce() -> T,
-) -> ExpectRevertReturn {
-    let success =
-        result.map(|(retdata, expected_revert)| retdata == expected_revert).unwrap_or(false);
-
-    match (success, call) {
-        // Success case for CALLs needs to return a dummy output value which
-        // can be decoded
-        (true, true) => ExpectRevertReturn::Call(Capture::Exit((
-            ExitReason::Succeed(ExitSucceed::Returned),
-            DUMMY_OUTPUT.to_vec(),
-        ))),
-        // Success case for CREATE doesn't need to return any value but must return a
-        // dummy address
-        (true, false) => ExpectRevertReturn::Create(Capture::Exit((
-            ExitReason::Succeed(ExitSucceed::Returned),
-            Some(Address::from_str("0000000000000000000000000000000000000001").unwrap()),
-            Vec::new(),
-        ))),
-        // Failure cases just return the abi encoded error
-        (false, true) => ExpectRevertReturn::Call(Capture::Exit((
-            ExitReason::Revert(ExitRevert::Reverted),
-            ethers::abi::encode(&[Token::String(err().to_string())]),
-        ))),
-        (false, false) => ExpectRevertReturn::Create(Capture::Exit((
-            ExitReason::Revert(ExitRevert::Reverted),
-            None,
-            ethers::abi::encode(&[Token::String(err().to_string())]),
-        ))),
     }
 }
